keep UID in long instead of int

+++ RoPP/ropp.h
@@ -26,12 +26,12 @@ namespace RoPP
             
 
 
-            User(int UID)
+            User(long UID)
             {
                 this->UID = UID;
             }
 
         private:
-            int UID;
+            long UID;
     };
 }
